Several features / bugs resolved:
https://www.pivotaltracker.com/story/show/154906423
Guests can - if they are admin or manager - invite members. To make
this work guests now can also be invited with the intended role
of manager and admin. As a consequence guests can now also delete
and resend invitations of they are admin or manager in a team.

https://www.pivotaltracker.com/story/show/154906454
The VOOT API endpoints accept fully qualified group names which are then
stripped to the local urn. The group name context - e.g. the urn prefix
which makes them unique - is configured as an ansible environment
variable.

https://www.pivotaltracker.com/story/show/152848011
If the Identity Provider does not provide a display name - provided
to teams as a 'displayName' shibboleth header - then a custom
error is shown which informs the user about this omission.